Makes the Mikado module executable
This helps with running/integrating mikado in other programs
Also allows for better support of IDE functionality in some cases
